define ac and gc_epoch as intptr_t, zero ac and gc_epoch at allocation

+++ src/hotspot/share/oops/oop.cpp
@@ -34,7 +34,7 @@
 #include "utilities/copy.hpp"
 
 bool always_do_update_barrier = false;
-int oopDesc::static_gc_epoch = 0;
+intptr_t oopDesc::static_gc_epoch = 0;
 
 void oopDesc::print_on(outputStream* st) const {
   if (this == NULL) {
